TNFSLib caching works, writing bypasses cache
Read caching has been added to TNFSlib.
**Please note that write caching has yet to be done,
so writes may actually be slower until that's completed**

Testing:
As currently implemented, the cache is entirely filled whenever it is
empty/emptied. This results in a delay whenever a "cache miss" occurs,
so there's a trade-off between cache size and the penalty incurred
every time there's a cache miss.

tnfslibMountInfo.h contains a #define for TNFS_FILE_CACHE_SIZE which
defines how big the cache is and how many bytes must be read every time
there's a cache miss and it needs to be filled.

Since Atari sector reads happen in 128-byte chunks, it's best to
set this to a multiple of 128 that fits within the payload for a TNFS
UDP packet. That payload size is calculated and definied in tnfslib.h
TNFS_MAX_READWRITE_PAYLOAD.  It's currently 501 bytes.

Since there is a lot of latency in receiving a UDP packet over the
Internet, it's best to keep the number of TNFS READ requests low.
Setting cache size to 3 sectors (3 * 128 = 384 bytes) will allow
the cache to be filled with a single TNFS READ request at the current
TNFS_MAX_READWRITE_PAYLOAD size.

If we set the cache size to 4 sectors (4 * 128 = 512), we'll need
to make two TNFS READ requests, since 512 is greater than
TNFS_MAX_READWRITE_PAYLOAD.  If you do this, you may as well increase
the cache size to what will fit in two TNFS READ requests.  That would
be 7 sectors (7 * 128 = 896).

Test different values of TNFS_FILE_CACHE_SIZE to observe the benefit of
a larger cache and the cost of a larger cache miss. From some quick
tests, a cache size of 3 or 7 sectors (one or two TNFS READ request)
seems to work best.

The other value that can be tested is TNFS_PAYLOAD_SIZE in tnfslib.h
This dictates the size of the UDP packet sent for every TNFS request.

The total size of the packet is:
(68 bytes IP+UDP header) + (4 bytes TNFS header) + TNFS_PAYLOAD_SIZE

The current value results in 576-byte packets, which are considered
safe and unlikely to be fragmented when sent across the Internet.

However, if we increase this size, we can fill a larger cache with a
single TNFS READ request, incurring less of the latency that multiple
requests would bring while still benefiting from a larger cache.

Increasing the TNFS_PAYLOAD_SIZE hasn't been tested yet. Any testing
should take into account the likelihood of the larger UDP packet being
fragmented when sent over the Internet.