Web UI - Show when drive slots rotated
Since rotation of disks (via [A] button) doesn't
remount what's in the Drive Slots, the new "Mounts" section
of the web UI can be misleading.

Now, if a drive slot doesn't match the device id, then it will
be shown.  e.g., normally the first is labeled "Drive Slot 1",
but if you've rotated once, it will be "Drive Slot 1 (D2:)".