Add multiple wifi slots
This fixes #493

The fuji will attempt to connect to the currently configured
wifi, and if that fails will attempt to match any networks
against stored entries, and try them in order of saved config
which means it will try the last one it was able to connect to
first if more than one network is found.

This means if you're constantly in range of multiple devices, you
can't currently pick the second one without removing the entry
for the other. This could be a future enhancement if needed.

Note it will only attempt to connect to something that is visible
from a network scan and ignore other saved entries (which massively
speeds up the process of connecting to one of the stored entries).

For new networks, as they are added, old networks are pushed down
the 'stack' of old wifi configs until a max is reached, at which
point they drop off the end. The pushed down networks are tried
again when the current cannot be connected to as detailed above.