RC2014: bus: implement PIN_PROCEED for data waiting
Currently, the RC2014 requires a command to check the status of each
device to determine when there is any data available. This causes a
flood of constant communication on the SPIO bus.

Fix this by implementing PIN_PROCEED which is asserted when there are
any (supported) devices that have data or change of status waiting. This
can then be polled by the Z80 host instead and only sending a command on
the SPI bus when required.