make sure interrupts always happen on core1
This patch moves the call to main_setup to inside the main service task.
This has the benefit of making sure interrupt service handlers are set
up to run on the same CPU core as the main service task, rather than on any available core.

This was done because of observed crashes on apple2 where the ISRs would
be running in core 0 and getting a wdt trying to obtain locks for
hardware being used at the same time in core 1.